moved config to /config folder instead of bento
